Remove Back-End and Front-End Detection Feature
Because it is impossible to detect front-end feature in the absence of a
back-end feature from the beginning, so the function of `index__.php`
file will never be different from the function of `index.php` file.

So it would be better if I remove both features and stick with the
generic `index.php` file.